requirements.txt
Django==4.1.3
djangorestframework==3.14.0
Pillow==9.3.0
django-filter==22.1